# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: TheChickenMan B2B Chicken Marketplace API
  version: 1.0.0
  description: Authoritative RESTful API for TheChickenMan commercial poultry platform.
servers:
  - url: /api/v1
    description: Production & Local API Gateway
paths:
...
  /buyers/profile:
    get:
      summary: Get current buyer profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Buyer profile }
    post:
      summary: Create or update buyer profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Buyer profile updated }
  /buyers/kyc:
    post:
      summary: Submit buyer KYC details
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: KYC submitted }
  /buyers/locations:
    get:
      summary: List buyer delivery locations
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of delivery locations }
    post:
      summary: Add a buyer delivery location
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Location added }
  /catalogue/products:
    get:
      summary: List standard chicken products
      responses:
        '200': { description: List of master SKUs }
    post:
      summary: Create standard chicken product (Admin)
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Master product created }
  /catalogue/listings:
    get:
      summary: Search and filter active supplier product listings
      parameters:
        - name: query
          in: query
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string, enum: [FRESH, CHILLED, FROZEN] }
        - name: pincode
          in: query
          schema: { type: string }
      responses:
        '200': { description: Active purchasable listings }
    post:
      summary: Create supplier SKU listing
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Listing created }
  /catalogue/listings/{id}:
    get:
      summary: Get product detail page with transparent landed pricing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Product listing details }
...
  /cart:
    get:
      summary: Get current buyer cart
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Cart with items and totals }
    post:
      summary: Add item to cart
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Item added }
    delete:
      summary: Clear cart or remove item
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Cart cleared }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT